Changed numbers to float.

+++ Prototype/src/prototype/Keywords.java
@@ -9,7 +9,7 @@ package prototype;
 //object for keyword array
 public class Keywords {
     public String keyword;
-    public int valence;
-    public int arousal;
-    public int dominance;
+    public float valence;
+    public float arousal;
+    public float dominance;
 }
